test(bridge): build full Playwright + Synpress v3 e2e harness
Expand the Playwright e2e foundation to cover the whole suite and remove the
shared Cypress harness:

- globalSetup ports the Cypress setupNodeEvents chain prep (ERC20/WETH deploy +
  fund, approvals, custom-gas-token, redeem-retryable tx) and fires the
  activity/assertion loops; globalSetup.cctp covers the CCTP testnet setup
- fixtures use a worker-scoped MetaMask context (set up once per worker) pinned
  to MetaMask 11.15.0, plus an e2eEnv config fixture
- support/common.ts and support/actions.ts port the network helpers and all
  Cypress custom commands as page-based async functions
- remove synpress.config.ts, synpress.cctp.config.ts, tests/support, and the
  Cypress-only config/types
