+++ packages/arb-token-bridge-ui/package.json
@@ -94,7 +94,6 @@
     "@types/react-virtualized": "^9.22.3",
     "@types/yargs": "^17.0.33",
     "autoprefixer": "^10.4.13",
-    "cypress-terminal-report": "^7.1.0",
     "env-cmd": "^10.1.0",
     "happy-dom": "^20.8.9",
     "jsdom-testing-mocks": "^1.16.0",
