--- packages/arb-token-bridge-ui/synpress-plugins.d.ts
@@ -1,1 +1,0 @@
-declare module '@synthetixio/synpress/plugins';
