--- packages/arb-token-bridge-ui/tests/e2e/README.md
@@ -1,19 +1,0 @@
-# End-to-end tests
-
-## Folder structure
-
-These folders hold end-to-end tests and supporting files for the Cypress Test Runner.
-
-- [fixtures](fixtures) holds optional JSON data for mocking, [read more](https://on.cypress.io/fixture)
-- [integration](integration) holds the actual test files, [read more](https://on.cypress.io/writing-and-organizing-tests)
-- [support](support) file runs before all tests and is a great place to write or load additional custom commands, [read more](https://on.cypress.io/writing-and-organizing-tests#Support-file)
-
-## `cypress.json` file
-
-You can configure project options in the [../cypress.json](../cypress.json) file, see [Cypress configuration doc](https://on.cypress.io/configuration).
-
-## More information
-
-- [https://github.com/cypress.io/cypress](https://github.com/cypress.io/cypress)
-- [https://docs.cypress.io/](https://docs.cypress.io/)
-- [Writing your first Cypress test](http://on.cypress.io/intro)
