--- packages/arb-token-bridge-ui/tests/e2e/cypress.d.ts
@@ -1,75 +1,0 @@
-/// <reference types="cypress" />
-import '@synthetixio/synpress/support/index.d.ts';
-import 'cypress-wait-until';
-
-import {
-  acceptTnC,
-  claimCctp,
-  clickClaimButton,
-  clickMoveFundsButton,
-  closeTransactionDetails,
-  confirmSpending,
-  fillCustomDestinationAddress,
-  findAmount2Input,
-  findAmountInput,
-  findClaimButton,
-  findDestinationChainButton,
-  findGasFeeSummary,
-  findMoveFundsButton,
-  findSelectTokenButton,
-  findSourceChainButton,
-  findTransactionDetailsCustomDestinationAddress,
-  findTransactionInTransactionHistory,
-  login,
-  openTransactionDetails,
-  searchAndSelectToken,
-  selectRoute,
-  selectTransactionsPanelTab,
-  switchToTransactionHistoryTab,
-  switchToTransferPanelTab,
-  typeAmount,
-  typeAmount2,
-} from '../support/commands';
-
-declare global {
-  namespace Cypress {
-    interface Chainable {
-      /**
-       * Custom command to connect MetaMask to the UI.
-       * @example cy.login()
-       */
-      acceptTnC(): typeof acceptTnC;
-      login: typeof login;
-      selectTransactionsPanelTab: typeof selectTransactionsPanelTab;
-      searchAndSelectToken({
-        tokenName,
-        tokenAddress,
-      }: {
-        tokenName: string;
-        tokenAddress: string;
-      }): typeof searchAndSelectToken;
-      fillCustomDestinationAddress(): typeof fillCustomDestinationAddress;
-      typeAmount: typeof typeAmount;
-      typeAmount2: typeof typeAmount2;
-      findAmountInput: typeof findAmountInput;
-      findAmount2Input: typeof findAmount2Input;
-      findSourceChainButton: typeof findSourceChainButton;
-      findDestinationChainButton: typeof findDestinationChainButton;
-      findGasFeeSummary: typeof findGasFeeSummary;
-      findMoveFundsButton: typeof findMoveFundsButton;
-      clickMoveFundsButton: typeof clickMoveFundsButton;
-      findSelectTokenButton: typeof findSelectTokenButton;
-      openTransactionDetails: typeof openTransactionDetails;
-      closeTransactionDetails: typeof closeTransactionDetails;
-      switchToTransferPanelTab: typeof switchToTransferPanelTab;
-      switchToTransactionHistoryTab: typeof switchToTransactionHistoryTab;
-      findTransactionDetailsCustomDestinationAddress: typeof findTransactionDetailsCustomDestinationAddress;
-      findTransactionInTransactionHistory: typeof findTransactionInTransactionHistory;
-      findClaimButton: typeof findClaimButton;
-      clickClaimButton: typeof clickClaimButton;
-      confirmSpending: typeof confirmSpending;
-      claimCctp: typeof claimCctp;
-      selectRoute: typeof selectRoute;
-    }
-  }
-}
