--- packages/arb-token-bridge-ui/tests/e2e/getCommonSynpressConfig.ts
@@ -1,22 +1,0 @@
-export function getCommonSynpressConfig(shouldRecordVideo: boolean) {
-  return {
-    userAgent: 'synpress',
-    retries: shouldRecordVideo ? 0 : 2,
-    screenshotsFolder: 'cypress/screenshots',
-    videosFolder: 'cypress/videos',
-    video: shouldRecordVideo,
-    screenshotOnRunFailure: true,
-    chromeWebSecurity: true,
-    modifyObstructiveCode: false,
-    scrollBehavior: false,
-    includeShadowDom: true,
-    viewportWidth: 1366,
-    viewportHeight: 850,
-    env: {
-      coverage: false,
-    },
-    defaultCommandTimeout: 30000,
-    pageLoadTimeout: 30000,
-    requestTimeout: 30000,
-  } as const satisfies Cypress.ConfigOptions;
-}
