--- packages/arb-token-bridge-ui/tests/support/common.ts
@@ -1,447 +1,0 @@
-/*
-  All the utility functions and configs related to our testing
-*/
-import { EthBridger, MultiCaller } from '@arbitrum/sdk';
-import { Provider, StaticJsonRpcProvider } from '@ethersproject/providers';
-import { BigNumber, Signer, Wallet, ethers, utils } from 'ethers';
-
-import { MULTICALL_TESTNET_ADDRESS } from '../../src/constants';
-import {
-  defaultL2Network,
-  defaultL3CustomGasTokenNetwork,
-  defaultL3Network,
-} from '../../src/util/networksNitroTestnode';
-
-export type NetworkType = 'parentChain' | 'childChain';
-export type NetworkName =
-  | 'Nitro Testnode L1'
-  | 'Nitro Testnode L2'
-  | 'Nitro Testnode L3'
-  | 'Arbitrum Sepolia'
-  | 'Ethereum'
-  | 'Sepolia';
-
-type NetworkConfig = {
-  networkName: NetworkName;
-  rpcUrl: string;
-  chainId: number;
-  symbol: string;
-  isTestnet: boolean;
-  multiCall: string;
-};
-
-export const getL1NetworkName = () => {
-  return getL1NetworkConfig().networkName;
-};
-
-export const getL2NetworkName = () => {
-  return getL2NetworkConfig().networkName;
-};
-
-export const getNetworkSlug = (network: 'parent' | 'child') => {
-  const networkName = network === 'parent' ? getL1NetworkName() : getL2NetworkName();
-  return networkName.toLowerCase().replace(' ', '-');
-};
-
-export const getL1NetworkConfig = (): NetworkConfig => {
-  const isOrbitTest = Cypress.env('ORBIT_TEST') == '1';
-
-  return {
-    networkName: isOrbitTest ? 'Nitro Testnode L2' : 'Nitro Testnode L1',
-    rpcUrl: Cypress.env('ETH_RPC_URL'),
-    chainId: isOrbitTest ? 412346 : 1337,
-    symbol: 'ETH',
-    isTestnet: true,
-    multiCall: isOrbitTest
-      ? defaultL2Network.tokenBridge!.childMultiCall
-      : defaultL2Network.tokenBridge!.parentMultiCall,
-  };
-};
-
-export const getL2NetworkConfig = (): NetworkConfig => {
-  const isOrbitTest = Cypress.env('ORBIT_TEST') == '1';
-  const nativeTokenSymbol = Cypress.env('NATIVE_TOKEN_SYMBOL') ?? 'ETH';
-  const isCustomFeeToken = nativeTokenSymbol !== 'ETH';
-
-  const l3Network = isCustomFeeToken ? defaultL3CustomGasTokenNetwork : defaultL3Network;
-
-  return {
-    networkName: isOrbitTest ? 'Nitro Testnode L3' : 'Nitro Testnode L2',
-    rpcUrl: Cypress.env('ARB_RPC_URL'),
-    chainId: isOrbitTest ? 333333 : 412346,
-    symbol: nativeTokenSymbol,
-    isTestnet: true,
-    multiCall: isOrbitTest
-      ? l3Network.tokenBridge!.childMultiCall
-      : defaultL2Network.tokenBridge!.childMultiCall,
-  };
-};
-
-export const getL1TestnetNetworkConfig = (): NetworkConfig => {
-  return {
-    networkName: 'Sepolia',
-    rpcUrl: Cypress.env('ETH_SEPOLIA_RPC_URL'),
-    chainId: 11155111,
-    symbol: 'ETH',
-    isTestnet: true,
-    multiCall: MULTICALL_TESTNET_ADDRESS,
-  };
-};
-
-export const getL2TestnetNetworkConfig = (): NetworkConfig => {
-  return {
-    networkName: 'Arbitrum Sepolia',
-    rpcUrl: Cypress.env('ARB_SEPOLIA_RPC_URL'),
-    chainId: 421614,
-    symbol: 'ETH',
-    isTestnet: true,
-    multiCall: MULTICALL_TESTNET_ADDRESS,
-  };
-};
-
-export const ERC20TokenName = 'Test Arbitrum Token';
-export const ERC20TokenSymbol = 'TESTARB';
-export const ERC20TokenDecimals = 18;
-export const invalidTokenAddress = utils.computeAddress(utils.randomBytes(32));
-
-export const moreThanZeroBalance = /0(\.\d+)/;
-
-export function getZeroToLessThanOneToken(symbol: string) {
-  return new RegExp(`0(\\.\\d+)* ${symbol}`);
-}
-
-export const importTokenThroughUI = (address: string) => {
-  // Click on the ETH dropdown (Select token button)
-  cy.findSelectTokenButton(Cypress.env('NATIVE_TOKEN_SYMBOL') ?? 'ETH').click();
-
-  // open the Select Token popup
-  cy.findByPlaceholderText(/Search by token name/i)
-    .should('be.visible')
-    .type(address);
-
-  // Click on the Add new token button
-  return cy.findByRole('button', { name: 'Add New Token' }).should('be.visible').click();
-};
-
-export async function getInitialETHBalance(
-  rpcURL: string,
-  walletAddress?: string,
-): Promise<BigNumber> {
-  const provider = new StaticJsonRpcProvider(rpcURL);
-  return await provider.getBalance(walletAddress ?? Cypress.env('ADDRESS'));
-}
-
-export async function getInitialERC20Balance({
-  tokenAddress,
-  multiCallerAddress,
-  rpcURL,
-  address,
-}: {
-  tokenAddress: string;
-  multiCallerAddress: string;
-  rpcURL: string;
-  address: string;
-}): Promise<BigNumber | undefined> {
-  const provider = new StaticJsonRpcProvider(rpcURL);
-  const multiCaller = new MultiCaller(provider, multiCallerAddress);
-  const [tokenData] = await multiCaller.getTokenData([tokenAddress], {
-    balanceOf: { account: address },
-  });
-  return tokenData?.balance;
-}
-
-export const acceptMetamaskAccess = () => {
-  cy.acceptMetamaskAccess().then(() => {
-    cy.isCypressWindowActive().then((cyWindowIsActive) => {
-      if (!cyWindowIsActive) {
-        cy.switchToCypressWindow().should('be.true');
-      }
-    });
-  });
-};
-
-export const startWebApp = (
-  url = '/bridge',
-  options: {
-    query: { [s: string]: string };
-    connectMetamask: boolean;
-  },
-) => {
-  // once all the metamask setup is done, we can start the actual web-app for testing
-  // clear local storage for terms to always have it pop up
-  cy.clearLocalStorage('arbitrum:bridge:tos-v2');
-  cy.visit(url, {
-    qs: options.query,
-  });
-  if (Cypress.currentRetry > 0) {
-    // ensures we don't test with the same state that could have caused the test to fail
-    cy.reload(true);
-  }
-
-  cy.task('getWalletConnectedToDapp').then((connected) => {
-    if (!connected) {
-      cy.findAllByText('Connect Wallet').first().should('be.visible');
-      if (options.connectMetamask) {
-        cy.findAllByText('Connect Wallet').first().click();
-        cy.contains('MetaMask').should('be.visible').click();
-        acceptMetamaskAccess();
-        cy.task('setWalletConnectedToDapp');
-      }
-    }
-  });
-};
-
-export const visitAfterSomeDelay = (url: string, options?: Partial<Cypress.VisitOptions>) => {
-  cy.wait(15_000); // let all the race conditions settle, let UI load well first
-  cy.visit(url, options);
-  cy.wait(15_000);
-};
-
-export const wait = (ms = 0): Promise<void> => {
-  return new Promise((res) => setTimeout(res, ms));
-};
-
-export async function generateActivityOnChains({
-  parentProvider,
-  childProvider,
-  wallet,
-}: {
-  parentProvider: Provider;
-  childProvider: Provider;
-  wallet: Wallet;
-}) {
-  const keepMining = async (miner: Signer) => {
-    /* eslint-disable no-await-in-loop */
-    while (true) {
-      await (
-        await miner.sendTransaction({
-          to: await miner.getAddress(),
-          value: 0,
-          // random data to make the tx heavy, so that batches are posted sooner (since they're posted according to calldata size)
-          data: '0x0000000000000000000000000000000000000000000000000000000000000020000000000000000000000000000000000000000000000000000000000000010c3c627574746f6e20636c6173733d226e61766261722d746f67676c65722220747970653d22627574746f6e2220646174612d746f67676c653d22636f6c6c617073652220646174612d7461726765743d22236e6176626172537570706f72746564436f6e74656e742220617269612d636f6e74726f6c733d226e6176626172537570706f72746564436f6e74656e742220617269612d657870616e6465643d2266616c73652220617269612d6c6162656c3d223c253d20676574746578742822546f67676c65206e617669676174696f6e222920253e223e203c7370616e20636c6173733d226e61766261722d746f67676c65722d69636f6e223e3c2f7370616e3e203c2f627574746f6e3e0000000000000000000000000000000000000000',
-        })
-      ).wait();
-
-      await wait(100);
-    }
-    /* eslint-enable no-await-in-loop */
-  };
-  // whilst waiting for status we mine on both parentChain and childChain
-  console.log('Generating activity on parentChain...');
-  const minerParent = Wallet.createRandom().connect(parentProvider);
-
-  const decimals = await getNativeTokenDecimals({
-    parentProvider,
-    childProvider,
-  });
-
-  await fundEth({
-    address: await minerParent.getAddress(),
-    provider: parentProvider,
-    sourceWallet: wallet,
-    networkType: 'parentChain',
-    amount: utils.parseUnits('0.2', decimals),
-  });
-
-  console.log('Generating activity on childChain...');
-  const minerChild = Wallet.createRandom().connect(childProvider);
-
-  await fundEth({
-    address: await minerChild.getAddress(),
-    provider: childProvider,
-    sourceWallet: wallet,
-    networkType: 'childChain',
-    amount: utils.parseEther('0.2'),
-  });
-
-  await Promise.allSettled([keepMining(minerParent), keepMining(minerChild)]);
-}
-
-export async function checkForAssertions({
-  parentProvider,
-  testType,
-}: {
-  parentProvider: Provider;
-  testType: 'regular' | 'orbit-eth' | 'orbit-custom';
-}) {
-  // Classic rollup ABI (pre-BoLD)
-  const classicAbi = [
-    'function latestConfirmed() public view returns (uint64)',
-    'function latestNodeCreated() public view returns (uint64)',
-  ];
-
-  // BoLD rollup ABI — latestConfirmed returns bytes32, no latestNodeCreated
-  const boldAbi = [
-    'function latestConfirmed() public view returns (bytes32)',
-    'function getAssertion(bytes32) public view returns (tuple(uint64 firstChildBlock, uint64 secondChildBlock, uint64 createdAtBlock, bool isFirstChild, uint8 status, bytes32 configHash))',
-    'event AssertionCreated(bytes32 indexed assertionHash, bytes32 indexed parentAssertionHash, tuple(tuple(bytes32 globalState, uint8 machineStatus) beforeState, tuple(bytes32 globalState, uint8 machineStatus) afterState) assertion, bytes32 afterInboxBatchAcc, uint256 inboxMaxCount, bytes32 wasmModuleRoot, uint256 requiredStake, address challengeManager, uint64 confirmPeriodBlocks)',
-  ];
-
-  let rollupAddress: string;
-
-  switch (testType) {
-    case 'orbit-eth':
-      rollupAddress = defaultL3Network.ethBridge.rollup;
-      break;
-    case 'orbit-custom':
-      rollupAddress = defaultL3CustomGasTokenNetwork.ethBridge.rollup;
-      break;
-    default:
-      rollupAddress = defaultL2Network.ethBridge.rollup;
-  }
-
-  const parentChainId = (await parentProvider.getNetwork()).chainId;
-
-  // Detect BoLD vs classic by trying the classic latestNodeCreated()
-  const classicContract = new ethers.Contract(rollupAddress, classicAbi, parentProvider);
-  let isBold = false;
-  try {
-    await classicContract.latestNodeCreated();
-  } catch {
-    isBold = true;
-    console.log(`Rollup ${rollupAddress} on ChainId ${parentChainId} is a BoLD rollup`);
-  }
-
-  if (isBold) {
-    const boldContract = new ethers.Contract(rollupAddress, boldAbi, parentProvider);
-    // For orbit tests the rollup lives on the Nitro L2, where `block.number` (and therefore
-    // `createdAtBlock`) is the L1 block number, so it can't be used as a block range on the L2
-    // directly. Map it via the NodeInterface precompile, like the SDK's
-    // ChildToParentMessageNitro.getBlockFromAssertionId does.
-    const nodeInterface = new ethers.Contract(
-      '0x00000000000000000000000000000000000000C8',
-      [
-        'function l2BlockRangeForL1(uint64 blockNum) view returns (uint64 firstBlock, uint64 lastBlock)',
-      ],
-      parentProvider,
-    );
-    /* eslint-disable no-await-in-loop */
-    while (true) {
-      try {
-        const latestConfirmed = await boldContract.latestConfirmed();
-        const assertion = await boldContract.getAssertion(latestConfirmed);
-        const createdAtBlock = assertion.createdAtBlock.toNumber();
-        const fromBlock =
-          testType === 'regular'
-            ? createdAtBlock
-            : (await nodeInterface.l2BlockRangeForL1(createdAtBlock)).firstBlock.toNumber();
-        // Count assertions created since the latest confirmed
-        const events = await boldContract.queryFilter('AssertionCreated', fromBlock, 'latest');
-        console.log(
-          `***** BoLD assertion status on ChainId ${parentChainId}: ${events.length} assertions since confirmed block ${fromBlock}, latest confirmed: ${latestConfirmed.slice(0, 10)}...`,
-        );
-      } catch (e) {
-        console.log(
-          `Could not fetch BoLD assertions for '${rollupAddress}' on ChainId ${parentChainId}`,
-          e,
-        );
-      }
-      await wait(10000);
-    }
-    /* eslint-enable no-await-in-loop */
-  } else {
-    /* eslint-disable no-await-in-loop */
-    while (true) {
-      try {
-        console.log(
-          `***** Assertion status on ChainId ${parentChainId}: ${(
-            await classicContract.latestNodeCreated()
-          ).toString()} created / ${(await classicContract.latestConfirmed()).toString()} confirmed`,
-        );
-      } catch (e) {
-        console.log(
-          `Could not fetch assertions for '${rollupAddress}' on ChainId ${parentChainId}`,
-          e,
-        );
-      }
-      await wait(10000);
-    }
-    /* eslint-enable no-await-in-loop */
-  }
-  /* eslint-enable no-await-in-loop */
-}
-
-export async function fundEth({
-  address, // wallet address where funding is required
-  provider,
-  sourceWallet, // source wallet that will fund the `address`,
-  networkType,
-  amount,
-}: {
-  address: string;
-  provider: Provider;
-  sourceWallet: Wallet;
-  networkType: NetworkType;
-  amount: BigNumber;
-}) {
-  console.log(`Funding ETH ${address} on ${networkType}...`);
-  const balance = await provider.getBalance(address);
-  // Fund only if the balance is less than 2 eth
-  if (balance.lt(amount)) {
-    const tx = await sourceWallet.connect(provider).sendTransaction({
-      to: address,
-      value: amount,
-    });
-    await tx.wait();
-  }
-}
-
-export async function getCustomDestinationAddress() {
-  console.log('Getting custom destination address...');
-  return (await Wallet.createRandom().getAddress()).toLowerCase();
-}
-
-export function setupCypressTasks(
-  on: Cypress.PluginEvents,
-  { requiresNetworkSetup }: { requiresNetworkSetup: boolean },
-) {
-  let currentNetworkName: NetworkName | null = null;
-  let networkSetupComplete = !requiresNetworkSetup;
-  let walletConnectedToDapp = false;
-
-  on('task', {
-    setCurrentNetworkName: (networkName: NetworkName) => {
-      currentNetworkName = networkName;
-      return null;
-    },
-    getCurrentNetworkName: () => {
-      return currentNetworkName;
-    },
-    setNetworkSetupComplete: () => {
-      networkSetupComplete = true;
-      return null;
-    },
-    getNetworkSetupComplete: () => {
-      return networkSetupComplete;
-    },
-    setWalletConnectedToDapp: () => {
-      walletConnectedToDapp = true;
-      return null;
-    },
-    getWalletConnectedToDapp: () => {
-      return walletConnectedToDapp;
-    },
-  });
-}
-
-export async function getNativeTokenDecimals({
-  parentProvider,
-  childProvider,
-}: {
-  parentProvider: Provider;
-  childProvider: Provider;
-}) {
-  const multiCaller = await MultiCaller.fromProvider(parentProvider);
-  const ethBridger = await EthBridger.fromProvider(childProvider);
-  const isCustomFeeToken = typeof ethBridger.nativeToken !== 'undefined';
-
-  const nativeToken = isCustomFeeToken
-    ? (
-        await multiCaller.getTokenData([ethBridger.nativeToken!], {
-          decimals: true,
-        })
-      )[0]
-    : undefined;
-
-  return nativeToken?.decimals ?? 18;
-}
